+++ ChangeLog.md
@@ -1,3 +1,9 @@
+2018-11-27 0.0.4
+
+- minor adjustments in data tables:
+    - missions: remove `name` from
+    - parameters: rename `name` to `paramName`
+
 2018-11-20 0.0.3
 
 - set `--classic-ids` so Postgraphile uses `id` (as mandated by the Relay spec) instead of `nodeId`